refactor(docker) Update repo structure to run tests workflow
