+++ Makefile
@@ -7,7 +7,7 @@ vet:	fmt
 	go vet ./...
 
 build: vet
-	go build -o ./bin/core-service 
+	go build -o ./bin/core-service ./cmd/main.go
 
 run: build
 	./bin/core-service
