+++ pkg/handlers/handlers.go
@@ -1,4 +1,4 @@
-package main
+package handlers
 
 import "github.com/gofiber/fiber/v2"
